+++ src/utils/translations.js
@@ -1,0 +1,438 @@
+const dict = {
+    // Abkhazian.
+    ab: false,
+    // Afar.
+    aa: false,
+    // Afrikaans.
+    af: {
+        Hide: "Versteek",
+    },
+    // Akan.
+    ak: false,
+    // Albanian.
+    sq: false,
+    // Amharic.
+    am: false,
+    // Arabic.
+    ar: {
+        Hide: "يخفي",
+    },
+    // Aragonese.
+    an: false,
+    // Armenian.
+    hy: false,
+    // Assamese.
+    as: false,
+    // Avaric.
+    av: false,
+    // Avestan.
+    ae: false,
+    // Aymara.
+    ay: false,
+    // Azerbaijani.
+    az: false,
+    // Bambara.
+    bm: false,
+    // Bashkir.
+    ba: false,
+    // Basque.
+    eu: false,
+    // Belarusian.
+    be: false,
+    // Bengali (Bangla).
+    bn: false,
+    // Bihari.
+    bh: false,
+    // Bislama.
+    bi: false,
+    // Bosnian.
+    bs: false,
+    // Breton.
+    br: false,
+    // Bulgarian.
+    bg: false,
+    // Burmese.
+    my: false,
+    // Catalan.
+    ca: false,
+    // Chamorro.
+    ch: false,
+    // Chechen.
+    ce: false,
+    // Chichewa, Chewa, Nyanja.
+    ny: false,
+    // Chinese.
+    zh: {
+        Hide: "隐藏",
+    },
+    // Chinese (Simplified).
+    "zh-Hans": false,
+    // Chinese (Traditional).
+    "zh-Hant": false,
+    // Chuvash.
+    cv: false,
+    // Cornish.
+    kw: false,
+    // Corsican.
+    co: false,
+    // Cree.
+    cr: false,
+    // Croatian.
+    hr: false,
+    // Czech.
+    cs: false,
+    // Danish.
+    da: false,
+    // Divehi, Dhivehi, Maldivian.
+    dv: false,
+    // Dutch.
+    nl: {
+        Hide: "Verbergen",
+    },
+    // Dzongkha.
+    dz: false,
+    // English.
+    en: {
+        Hide: "Hide",
+    },
+    // Esperanto.
+    eo: false,
+    // Estonian.
+    et: false,
+    // Ewe.
+    ee: false,
+    // Faroese.
+    fo: false,
+    // Fijian.
+    fj: false,
+    // Finnish.
+    fi: false,
+    // French.
+    fr: {
+        Hide: "Cacher",
+    },
+    // Fula, Fulah, Pulaar, Pular.
+    ff: false,
+    // Galician.
+    gl: false,
+    // Gaelic (Scottish).
+    gd: false,
+    // Gaelic (Manx).
+    gv: false,
+    // Georgian.
+    ka: false,
+    // German.
+    de: {
+        Hide: "Verstecken",
+    },
+    // Greek.
+    el: false,
+    // Greenlandic.
+    kl: false,
+    // Guarani.
+    gn: false,
+    // Gujarati.
+    gu: false,
+    // Haitian Creole.
+    ht: false,
+    // Hausa.
+    ha: false,
+    // Hebrew.
+    he: false,
+    // Herero.
+    hz: false,
+    // Hindi.
+    hi: {
+        Hide: "छिपाना",
+    },
+    // Hiri Motu.
+    ho: false,
+    // Hungarian.
+    hu: false,
+    // Icelandic.
+    is: false,
+    // Ido.
+    io: false,
+    // Igbo.
+    ig: false,
+    // Indonesian.
+    id: false,
+    // Indonesian.
+    in: false,
+    // Interlingua.
+    ia: false,
+    // Interlingue.
+    ie: false,
+    // Inuktitut.
+    iu: false,
+    // Inupiak.
+    ik: false,
+    // Irish.
+    ga: false,
+    // Italian.
+    it: {
+        Hide: "Nascondere",
+    },
+    // Japanese.
+    ja: {
+        Hide: "隠れる",
+    },
+    // Javanese.
+    jv: false,
+    // Kalaallisut, Greenlandic.
+    kl: false,
+    // Kannada.
+    kn: false,
+    // Kanuri.
+    kr: false,
+    // Kashmiri.
+    ks: false,
+    // Kazakh.
+    kk: false,
+    // Khmer.
+    km: false,
+    // Kikuyu.
+    ki: false,
+    // Kinyarwanda (Rwanda).
+    rw: false,
+    // Kirundi.
+    rn: false,
+    // Kyrgyz.
+    ky: false,
+    // Komi.
+    kv: false,
+    // Kongo.
+    kg: false,
+    // Korean.
+    ko: {
+        Hide: "숨다",
+    },
+    // Kurdish.
+    ku: false,
+    // Kwanyama.
+    kj: false,
+    // Lao.
+    lo: false,
+    // Latin.
+    la: false,
+    // Latvian (Lettish).
+    lv: false,
+    // Limburgish ( Limburger).
+    li: false,
+    // Lingala.
+    ln: false,
+    // Lithuanian.
+    lt: false,
+    // Luga-Katanga.
+    lu: false,
+    // Luganda, Ganda.
+    lg: false,
+    // Luxembourgish.
+    lb: false,
+    // Manx.
+    gv: false,
+    // Macedonian.
+    mk: false,
+    // Malagasy.
+    mg: false,
+    // Malay.
+    ms: false,
+    // Malayalam.
+    ml: false,
+    // Maltese.
+    mt: false,
+    // Maori.
+    mi: false,
+    // Marathi.
+    mr: false,
+    // Marshallese.
+    mh: false,
+    // Moldavian.
+    mo: false,
+    // Mongolian.
+    mn: false,
+    // Nauru.
+    na: false,
+    // Navajo.
+    nv: false,
+    // Ndonga.
+    ng: false,
+    // Northern Ndebele.
+    nd: false,
+    // Nepali.
+    ne: false,
+    // Norwegian.
+    no: false,
+    // Norwegian bokmål.
+    nb: false,
+    // Norwegian nynorsk.
+    nn: false,
+    // Nuosu.
+    ii: false,
+    // Occitan.
+    oc: false,
+    // Ojibwe.
+    oj: false,
+    // Old Church Slavonic, Old Bulgarian.
+    cu: false,
+    // Oriya.
+    or: false,
+    // Oromo (Afaan Oromo).
+    om: false,
+    // Ossetian.
+    os: false,
+    // Pāli.
+    pi: false,
+    // Pashto, Pushto.
+    ps: false,
+    // Persian (Farsi).
+    fa: false,
+    // Polish.
+    pl: false,
+    // Portuguese.
+    pt: false,
+    // Punjabi (Eastern).
+    pa: false,
+    // Quechua.
+    qu: false,
+    // Romansh.
+    rm: false,
+    // Romanian.
+    ro: false,
+    // Russian.
+    ru: {
+        Hide: "Скрывать",
+    },
+    // Sami.
+    se: false,
+    // Samoan.
+    sm: false,
+    // Sango.
+    sg: false,
+    // Sanskrit.
+    sa: false,
+    // Serbian.
+    sr: false,
+    // Serbo-Croatian.
+    sh: false,
+    // Sesotho.
+    st: false,
+    // Setswana.
+    tn: false,
+    // Shona.
+    sn: false,
+    // Sichuan Yi.
+    ii: false,
+    // Sindhi.
+    sd: false,
+    // Sinhalese.
+    si: false,
+    // Siswati.
+    ss: false,
+    // Slovak.
+    sk: false,
+    // Slovenian.
+    sl: false,
+    // Somali.
+    so: false,
+    // Southern Ndebele.
+    nr: false,
+    // Spanish.
+    es: {
+        Hide: "Esconder",
+    },
+    // Sundanese.
+    su: false,
+    // Swahili (Kiswahili).
+    sw: false,
+    // Swati.
+    ss: false,
+    // Swedish.
+    sv: false,
+    // Tagalog.
+    tl: false,
+    // Tahitian.
+    ty: false,
+    // Tajik.
+    tg: false,
+    // Tamil.
+    ta: false,
+    // Tatar.
+    tt: false,
+    // Telugu.
+    te: false,
+    // Thai.
+    th: {
+        Hide: "ซ่อน",
+    },
+    // Tibetan.
+    bo: false,
+    // Tigrinya.
+    ti: false,
+    // Tonga.
+    to: false,
+    // Tsonga.
+    ts: false,
+    // Turkish.
+    tr: false,
+    // Turkmen.
+    tk: false,
+    // Twi.
+    tw: false,
+    // Uyghur.
+    ug: false,
+    // Ukrainian.
+    uk: false,
+    // Urdu.
+    ur: false,
+    // Uzbek.
+    uz: false,
+    // Venda.
+    ve: false,
+    // Vietnamese.
+    vi: false,
+    // Volapük.
+    vo: false,
+    // Wallon.
+    wa: false,
+    // Welsh.
+    cy: false,
+    // Wolof.
+    wo: false,
+    // Western Frisian.
+    fy: false,
+    // Xhosa.
+    xh: false,
+    // Yiddish.
+    yi: false,
+    // Yiddish.
+    ji: false,
+    // Yoruba.
+    yo: false,
+    // Zhuang, Chuang.
+    za: false,
+    // Zulu.
+    zu: false,
+};
+const events = [];
+let listening = false;
+export default function main(str) {
+    if (!listening) {
+        window.addEventListener("languagechange", handleLanguageChange);
+        listening = true;
+    }
+    const userLanguagePreference = navigator.language || "en";
+    if (!dict[userLanguagePreference]) {
+        return dict.en[str] || str;
+    }
+    return dict[userLanguagePreference][str] || str;
+}
+main.subscribe = function (callback) {
+    events.push(callback);
+};
+function handleLanguageChange() {
+    const newLang = navigator.language;
+    if (dict[newLang]) {
+        events.forEach((callback) => callback());
+    }
+}